# ORC-1122: [C++] Add buffer to decode the whole run in RleDecoderV2