@stiga-huang 2022-03-11
Tested with random unsigned numbers fitting into bitSize n with a given ratio of nulls. Reading 1B rows for each test.

| bitSize | % nulls | baseline avg(s) | optimized avg(s) | Speedup      |
|---------|---------|-----------------|------------------|--------------|
| 4       | 0%      | 3.52            | 2.745            | 1.2823315118 |
| 4       | 5%      | 7.925           | 7.185            | 1.1029923452 |
| 4       | 10%     | 8.48            | 6.805            | 1.2461425422 |
| 4       | 25%     | 11.15           | 6.475            | 1.722007722  |
| 4       | 50%     | 11.135          | 6.47             | 1.7210200927 |
| 8       | 0%      | 3.51            | 2.73             | 1.2857142857 |
| 8       | 5%      | 7.89            | 7.13             | 1.1065918654 |
| 8       | 10%     | 8.47            | 6.795            | 1.2465047829 |
| 8       | 25%     | 9.93            | 6.44             | 1.5419254658 |
| 8       | 50%     | 11.145          | 6.465            | 1.7238979118 |
| 16      | 0%      | 3.53            | 2.745            | 1.2859744991 |
| 16      | 5%      | 7.925           | 7.155            | 1.107617051  |
| 16      | 10%     | 8.46            | 6.8              | 1.2441176471 |
| 16      | 25%     | 9.92            | 6.41             | 1.5475819033 |
| 16      | 50%     | 11.2            | 6.47             | 1.7310664606 |
| 32      | 0%      | 3.515           | 2.745            | 1.2805100182 |
| 32      | 5%      | 7.915           | 7.15             | 1.106993007  |
| 32      | 10%     | 8.475           | 6.815            | 1.2435803375 |
| 32      | 25%     | 9.935           | 6.48             | 1.5331790123 |
| 32      | 50%     | 11.175          | 6.485            | 1.7232074017 |
| 32      | 100%    | 2.19            | 2.145            | 1.020979021  |

We can see speedup in all cases.

Tested on Ubuntu16.04 with a 6 cores CPU (12 virtual cores) and 32GB RAM.
CPU: Intel(R) Core(TM) i7-8700 CPU @ 3.20GHz