@dongjoon-hyun 2022-03-11
BTW, I copied your benchmark result into the PR description. When you click `Squash and merge` button, please copy and paste the PR description into the edit box in order to make a up-to-date commit log correctly.